Move rules to live within server as they are blitz specific business concerns and depend on next.js
