+++ packages/server/src/dev.ts
@@ -2,7 +2,7 @@ import {resolve} from 'path'
 import {synchronizeFiles as defaultSynchronizer} from '@blitzjs/synchronizer'
 import {ServerConfig, enhance} from './config'
 import {nextStartDev} from './next-utils'
-
+import {rules} from './rules'
 export async function dev(config: ServerConfig, readyForNextDev: Promise<any> = Promise.resolve()) {
   const {
     rootFolder,
@@ -23,6 +23,7 @@ export async function dev(config: ServerConfig, readyForNextDev: Promise<any> = 
 
   const [{manifest}] = await Promise.all([
     synchronizeFiles({
+      rules,
       src,
       dest,
       watch,
