+++ packages/server/src/next-utils.ts
@@ -1,6 +1,6 @@
 import {spawn} from 'cross-spawn'
 import detect from 'detect-port'
-import {Manifest} from '@blitzjs/synchronizer'
+import {Manifest} from './rules/manifest'
 import {through} from './streams'
 import {ServerConfig} from 'config'
