+++ packages/server/src/rules/config/index.ts
@@ -2,8 +2,8 @@ import {pathExistsSync} from 'fs-extra'
 import {resolve} from 'path'
 import File from 'vinyl'
 
-import {through} from '../../../streams'
-import {Rule} from '../../../types'
+import {through} from '../../streams'
+import {Rule} from '@blitzjs/synchronizer'
 
 const isNextConfigPath = (p: string) => /next\.config\.(js|ts)/.test(p)
 const isNowBuild = () => process.env.NOW_BUILDER || process.env.VERCEL_BUILDER
