+++ packages/server/src/rules/manifest/manifest.test.ts
