+++ packages/server/src/rules/utils.ts
@@ -1,4 +1,3 @@
-import File from 'vinyl'
 import {relative, resolve} from 'path'
 
 // Transform an absolute path with a relative path transformer
@@ -8,10 +7,6 @@ export const absolutePathTransform = (sourceFolder = '') => (relativeTransformer
   const startingPath = relative(sourceFolder, filePath)
   const transformedPath = relativeTransformer(startingPath)
   return resolve(sourceFolder, transformedPath)
-}
-
-export function isSourceFile(file: File) {
-  return file.hash.indexOf(':') === -1
 }
 
 export const getDuplicatePaths = (entries: string[], type: string) => {
